refactor import / module structure, add AggregatorStrategy enum
Change-type: patch
