+++ src/enums.ts
@@ -1,0 +1,10 @@
+export enum AggregatorStrategy {
+	SUM = 'sum',
+	FIRST = 'first',
+	MIN = 'min',
+	MAX = 'max',
+	AVERAGE = 'average',
+	OMIT = 'omit',
+}
+
+
